Feature/rebase master

+++ src/utils/StorageManager/migrations/addReuseCountMigration.ts
@@ -5,10 +5,14 @@ export class AddReuseCountColumn163837490194410 implements MigrationInterface {
         const tableExists = await queryRunner.hasTable('IdentityVerificationStorage');
 
         if (tableExists) {
-            await queryRunner.query(`
-            ALTER TABLE "IdentityVerificationStorage" 
-            ADD COLUMN "reuseCount" integer NOT NULL DEFAULT 0
+            const columnExists = await queryRunner.hasColumn('IdentityVerificationStorage', 'reuseCount');
+
+            if (!columnExists) {
+                await queryRunner.query(`
+                ALTER TABLE "IdentityVerificationStorage" 
+                ADD COLUMN "reuseCount" integer NOT NULL DEFAULT 0
             `);
+            }
         }
     }
 
@@ -16,10 +20,14 @@ export class AddReuseCountColumn163837490194410 implements MigrationInterface {
         const tableExists = await queryRunner.hasTable('IdentityVerificationStorage');
 
         if (tableExists) {
-            await queryRunner.query(`
-                ALTER TABLE "IdentityVerificationStorage" 
-                DROP COLUMN "reuseCount"
-            `);
+            const columnExists = await queryRunner.hasColumn('IdentityVerificationStorage', 'reuseCount');
+
+            if (columnExists) {
+                await queryRunner.query(`
+                    ALTER TABLE "IdentityVerificationStorage" 
+                    DROP COLUMN "reuseCount"
+                `);
+            }
         }
     }
 }
